+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.0.9] - 2020-03-16
+
+### Added
+- **`maximumStoredEvents`** field in `PeachCollector` to limit queue size
+- **`maximumStorageDays`**  field in `PeachCollector` to limit queue size
 
 ## [1.0.8] - 2020-02-07